requirements.txt
atomicwrites
attrs
beautifulsoup4
bs4
colorama
comtypes
glob2
importlib-metadata
iniconfig
jsonpickle
Mako
MarkupSafe
more-itertools
packaging
parse
parse-type
pluggy
pony
PyMySQL
pyparsing
pytest
pytest-bdd
pywin32
pywinauto
selenium
six
soupsieve
toml
urllib3
zipp
lxml
allure-pytest